requirements.txt
amqp==5.2.0
annotated-types==0.7.0
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.1
black==24.10.0
blessed==1.20.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contextlib2==21.6.0
cryptography==43.0.3
defusedxml==0.8.0rc2
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.5.0
django-ninja==1.3.0
django-ninja-extra==0.21.4
django-ninja-jwt==5.3.4
django-picklefield==3.2
django-q2==1.7.3
httplib2==0.22.0
idna==3.10
injector==0.22.0
jinxed==1.3.0
kombu==5.4.2
mypy-extensions==1.0.0
oauth2==1.9.0.post1
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
pika==1.3.2
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-openid==2.2.5
python3-openid==3.2.0
pytz==2024.2
redis==3.5.3
requests==2.32.3
requests-oauthlib==1.3.1
setuptools==75.3.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tweepy==4.14.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
